Add NK_login_enum to the C API
As with the C++ API, identifying the model by the first character of a
string is not intuitive.  This patch adds the NK_login_enum function to
the C API that accepts an enum value instead of a string.  It also adds
a new enum NK_device_model as device::DeviceModel is an enum class and
therefore cannot be used in C.