Correct return value for NitrokeyManager::connect()
NitrokeyManager::connect() currently returns true if the device pointer
is set.  Yet this does not mean that the connection was successful.  For
example, NitrokeyManger::connect(const char*) sets the device pointer
even if it was not successful.

This patch introduces a variable that keeps track of the connection
instead of checking the device pointer.  This corrects the return value
without changing the behavior of the connect method (returning the
Storage device if both a Pro and a Storage device are present).