Expose device model in C API
The C++ API already provides access to the model of the connected device
in NitrokeyManager::get_connected_device_model().  This patch also
exposes this information in the C API by adding NK_get_device_model.  As
there might be no device connected, the function returns a boolean
indicating the connection status and writes the model of the connected
device to a pointer passed as an argument.