Add getter for Storage status to C API
The C++ API currently provides the Storage status in
NitrokeyManager::get_status_storage().  The C API only provides a string
version of this data (NK_get_status_storage_as_string).  This patch adds
a struct to the C API that can store the storage status and a function
that can retrieve it.

The interpretation of the fields of the internal struct is based on the
following code in the Nitrokey Storage firmware:
    - src/HighLevelFunctions/FlashStorage.h, lines 73 to 90 (struct
      definition with comments)
    - src/OTP/report_protocol.c, lines 241 to 376 (debug output of the
      data)