Fix unit test for get_library_version for release commits
`git describe` may return a tag name or <tag>-<n>-g<hash>, where <n> is
the number of commits since the last tag <tag> and <hash> is the hash of
the current commit.  The current test case only considers the latter
case.  This patch adds a regular expression to handle both cases.