Add a test to ensure valid C code in the C API
All current tests are written in C++.  Therefore C++-only code in the C
API might not be detected by the tests.  This patch adds a minimal test
written in C that only includes the C API header.  This should make sure
that the C API code is valid.

The new test is called `test_minimal.c` and is added to the CMake
COMPILE_OFFLINE_TESTS option.