Change NitrokeyManager::connect_with_path to also work with Pro
Previously, Stick20 was hardcoded in connect_with_path.  Now we first
use hid_enumerate to find out the model on that path, then we connect to
that model.

We also could have added the model as a parameter to connect_with_path.
Yet we cannot directly check the model after connecting, so this would
be error-prone.