Return -2 if an error occured in NK_get_progress_bar_value
NK_get_progress_bar_value returns the progress value from 0 to 100 or -1
if there is no pending operation.  In the previous implementation, it
also returned zero if an error occurred, making it impossible to
distinguish progress zero and an error.  Therefore, we change the return
value to -2 if an error occured.