Add NK_get_status function to get status struct
Currently, the C API provides access to the device status using the
NK_get_{major,minor}_firmware_version, NK_device_serial_number and
NK_read_config functions.  Each function sends a command to the device,
although the data is returned in a single response.  This patch adds a
NK_status struct and a NK_get_status function that provide access to
this data with a single command.