test_pro: Extend test_get_status test case
This patch extends the test_get_status test case to check the content of
the returned status struct.  The firmware version, serial number and
config should match the values returned by the NK_*_firmware_version,
NK_device_serial_number and NK_read_config functions.

This does not work for the Nitrokey Storage, which will be fixed in the
next patch.