Also execute GET_DEVICE_STATUS in get_status for Storage
As described in this issue [0], the GET_STATUS command does not work for
the Nitrokey Storage.  Therefore, NitrokeyManager::get_status does not
work either.  To fix this, we additionally execute the GET_DEVICE_STATUS
command if a Nitrokey Storage is detected to determine the serial number
and the firmware version.  This also fixes the failing test
test_get_status that we introduced in the last patch.

Fixes #166.

[0] https://github.com/Nitrokey/nitrokey-storage-firmware/issues/96