Return serial number as uint32_t from C API
This patch adds the function NK_device_serial_number_as_u32 to the C
API.  It is similar to NK_device_serial_number but returns the raw
unsigned integer instead of a formatted string.  This patch also adds a
simple test case that ensures that the number is not zero.

Fixes #172.