Fix device discovery in NitrokeyManager::connect_with_path
The device discovery in NitrokeyManager's connect_with_path method is
broken. It tries to statically determine the vendor ID by using the
first working HID pointer. That's obviously not working when a both a
Nitrokey device and Librem Key are connected and the path is describing
a Librem Key. In such a case the vendor ID used would be that for the
Nitrokey and we would not be able to find the Librem Key at all. This
change fixes up the logic.